adjusted dependencies in Cargo files -- migrated functions from HolonSpace to HolonSpaceManager
