+++ crates/coordinator/core_schema/Cargo.toml
@@ -3,33 +3,37 @@ name = "core_schema"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
 crate-type = ["cdylib", "rlib"]
 name = "core_schema"
 
 [dependencies]
-
 # General Dependencies
 serde = { workspace = true }
 derive-new = { workspace = true }
 thiserror = { workspace = true }
-Inflector = {workspace = true}
-strum = {workspace = true}
-strum_macros ={workspace = true}
+Inflector = { workspace = true }
+strum = { workspace = true }
+strum_macros = { workspace = true }
 
 # Holochain Dependencies
 hdi = { workspace = true }
 hdk = { workspace = true }
-#holochain_integrity_types = { workspace = true }
 
 # MAP Dependencies
-
 holons_integrity = { workspace = true }
-holons = {workspace = true}
-shared_types_holon = {workspace = true}
-descriptors = {workspace = true}
+holons = { workspace = true }
+shared_types_holon = { workspace = true }
+descriptors = { workspace = true }
 
+# General dev-dependencies (platform-agnostic)
+[dev-dependencies]
+async-std = { workspace = true }
+futures = { workspace = true }
+pretty_assertions = { workspace = true }
+rstest = { workspace = true }
 
-
+# Target-specific dev dependencies for non-WASM builds
+[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
+holochain = { version = "0.3", default-features = false, features = ["test_utils", "sqlite-encrypted"] }
+tokio = { version = "1.35.1", features = ["full"] }
