+++ crates/coordinator/dances/Cargo.toml
@@ -11,26 +11,23 @@ name = "dances"
 hdk = { workspace = true }
 hdi = { workspace = true }
 holochain_integrity_types = { workspace = true }
-thiserror = { workspace = true, features = [] }
-#holochain = "0.1.7"
-
-#holons_integrity = { workspace = true }
+thiserror = { workspace = true }
 serde = { workspace = true }
 derive-new = { workspace = true }
 
 shared_types_holon = { workspace = true }
-holons = {workspace = true}
-descriptors = {workspace = true}
-core_schema = {workspace = true}
+holons = { workspace = true }
+descriptors = { workspace = true }
+core_schema = { workspace = true }
 
+# General dev-dependencies (platform-agnostic)
+[dev-dependencies]
+async-std = { workspace = true }
+futures = { workspace = true }
+pretty_assertions = { workspace = true }
+rstest = { workspace = true }
+
+# Target-specific dev dependencies for non-WASM builds
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-
-holochain = { version = "0.3", default-features = false, features = ["test_utils","sqlite-encrypted"] }
+holochain = { version = "0.3", default-features = false, features = ["test_utils", "sqlite-encrypted"] }
 tokio = { version = "1.35.1", features = ["full"] }
-futures = { version = "0.3.1", default-features = false }
-rstest = "0.21"
-async-std = "1.1.12"
-
-pretty_assertions = "1.4.0"
-
-#[dev-dependencies]
