+++ crates/coordinator/descriptors/Cargo.toml
@@ -7,10 +7,6 @@ edition = "2021"
 crate-type = ["cdylib", "rlib"]
 name = "descriptors"
 
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-
 [dependencies]
 # General Dependencies
 serde = { workspace = true }
@@ -19,24 +15,23 @@ thiserror = { workspace = true }
 
 # Holochain Dependencies
 hdk = { workspace = true }
-hdi = { workspace = true}
+hdi = { workspace = true }
 holochain_integrity_types = { workspace = true }
 
 # MAP Dependencies
-
 holons_integrity = { workspace = true }
-holons = {workspace = true}
-shared_types_holon = {workspace = true}
+holons = { workspace = true }
+shared_types_holon = { workspace = true }
+Inflector = { workspace = true }
 
-Inflector = {workspace = true}
+# General dev-dependencies (platform-agnostic)
+[dev-dependencies]
+async-std = { workspace = true }
+futures = { workspace = true }
+pretty_assertions = { workspace = true }
+rstest = { workspace = true }
 
+# Target-specific dev dependencies for non-WASM builds
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-
-holochain = {version = "0.3", default-features = false, features = ["test_utils","sqlite-encrypted"] }
-tokio = { version = "1.3", features = ["full"] }
-futures = { version = "0.3.1", default-features = false }
-rstest = "0.21"
-async-std = "1.1.12"
-
-#[dev-dependencies]
-
+holochain = { version = "0.3", default-features = false, features = ["test_utils", "sqlite-encrypted"] }
+tokio = { version = "1.35.1", features = ["full"] }
