+++ crates/coordinator/holons/Cargo.toml
@@ -12,25 +12,22 @@ derive-new = { workspace = true }
 hdk = { workspace = true }
 hdi = { workspace = true }
 holochain_integrity_types = { workspace = true }
-
 holons_integrity = { workspace = true }
-
 quick_cache = { workspace = true }
 thiserror = { workspace = true }
-
 shared_types_holon = { workspace = true }
 shared_validation = { workspace = true }
-
 serde = { workspace = true }
 serde_json = { workspace = true }
 
+[dev-dependencies]
+# General dev-dependencies (platform-agnostic)
+async-std = { workspace = true }
+futures = { workspace = true }
+pretty_assertions = { workspace = true }
+rstest = { workspace = true }
 
+# Target-specific dev dependencies for non-WASM builds
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-async-std = "1.1.12"
-futures = { version = "0.3.1", default-features = false }
-holochain = { version = "0.3", default-features = false, features = ["test_utils","sqlite-encrypted"] }
-pretty_assertions = "1.4.0"
-rstest = "0.21"
+holochain = { version = "0.3", default-features = false, features = ["test_utils", "sqlite-encrypted"] }
 tokio = { version = "1.35.1", features = ["full"] }
-
-#[dev-dependencies]
